+++ test/typescript-tests/testTypes.ts
@@ -1,3 +1,5 @@
+import * as assert from 'assert'
+import { expectTypeOf } from 'expect-type'
 import {
   AccessorNode,
   addDependencies,
@@ -21,15 +23,19 @@ import {
   Help,
   Index,
   IndexNode,
+  isSymbolNode,
   LUDecomposition,
   MathArray,
   MathCollection,
   MathJsChain,
   MathJsFunctionName,
   MathNode,
+  MathNodeCommon,
   MathNumericType,
+  MathScalarType,
   MathType,
   Matrix,
+  Node,
   ObjectNode,
   OperatorNode,
   OperatorNodeFn,
@@ -41,15 +47,9 @@ import {
   SimplifyRule,
   SLUDecomposition,
   SymbolNode,
-  MathNodeCommon,
   Unit,
-  UnitPrefix,
-  Node,
-  isSymbolNode,
-  MathScalarType
+  UnitPrefix
 } from 'mathjs'
-import * as assert from 'assert'
-import { expectTypeOf } from 'expect-type'
 
 // This file serves a dual purpose:
 // 1) examples of how to use math.js in TypeScript
@@ -649,6 +649,9 @@ Chaining examples
   expectTypeOf(math.chain([1]).ceil()).toMatchTypeOf<
     MathJsChain<MathCollection>
   >()
+  expectTypeOf(
+    math.chain(math.unit('5.2cm')).ceil(math.unit('cm'))
+  ).toMatchTypeOf<MathJsChain<Unit>>()
 
   // fix
   expectTypeOf(math.chain(1).fix()).toMatchTypeOf<
@@ -657,6 +660,9 @@ Chaining examples
   expectTypeOf(math.chain([1]).fix()).toMatchTypeOf<
     MathJsChain<MathCollection>
   >()
+  expectTypeOf(
+    math.chain(math.unit('5.2cm')).fix(math.unit('cm'))
+  ).toMatchTypeOf<MathJsChain<Unit>>()
 
   // floor
   expectTypeOf(math.chain(1).floor()).toMatchTypeOf<
@@ -665,6 +671,12 @@ Chaining examples
   expectTypeOf(math.chain([1]).floor()).toMatchTypeOf<
     MathJsChain<MathCollection>
   >()
+  expectTypeOf(
+    math.chain(math.unit('5.2cm')).floor(math.unit('cm'))
+  ).toMatchTypeOf<MathJsChain<Unit>>()
+  expectTypeOf(
+    math.chain(math.unit('5.2cm')).round(2, math.unit('cm'))
+  ).toMatchTypeOf<MathJsChain<Unit>>()
 
   // round
   expectTypeOf(math.chain(1).round()).toMatchTypeOf<
@@ -845,7 +857,6 @@ Chaining examples
 
   assert.throws(
     () =>
-      // @ts-expect-error ... gcd() supports only 1d matrices!
       math.gcd([
         [
           [1, 5],
@@ -1292,8 +1303,20 @@ Matrices examples
 
   // map matrix
   {
+    const arr = [1, 2, 3]
     assert.deepStrictEqual(
-      math.map([1, 2, 3], function (value) {
+      math.map(arr, function (value) {
+        return value * value
+      }),
+      [1, 4, 9]
+    )
+
+    assert.deepStrictEqual(
+      math.map(arr, function (value, index, self) {
+        const indexValue = index[0]
+        expectTypeOf(indexValue).toMatchTypeOf<number>()
+        assert.deepStrictEqual(self, arr)
+
         return value * value
       }),
       [1, 4, 9]
@@ -1302,8 +1325,9 @@ Matrices examples
 
   // filter matrix
   {
+    const arr = [6, -2, -1, 4, 3]
     assert.deepStrictEqual(
-      math.filter([6, -2, -1, 4, 3], function (x) {
+      math.filter(arr, function (x) {
         return x > 0
       }),
       [6, 4, 3]
@@ -1312,6 +1336,31 @@ Matrices examples
       math.filter(['23', 'foo', '100', '55', 'bar'], /\d+/),
       ['23', '100', '55']
     )
+    assert.deepStrictEqual(
+      math.filter(arr, function (x, index, self) {
+        const indexValue = index[0]
+        expectTypeOf(indexValue).toMatchTypeOf<number>()
+        assert.deepStrictEqual(self, arr)
+
+        return x > 0
+      }),
+      [6, 4, 3]
+    )
+  }
+
+  // forEach matrix
+  {
+    const arr = [6, -2, -1, 4, 3]
+
+    const output: number[] = []
+    math.forEach(arr, function (x, index, self) {
+      const indexValue = index[0]
+      expectTypeOf(indexValue).toMatchTypeOf<number>()
+      assert.deepStrictEqual(self, arr)
+      output.push(x)
+    })
+
+    assert.deepStrictEqual(output, arr)
   }
 
   // concat matrix
@@ -1450,6 +1499,9 @@ Math types examples: Type results after multiplying  'MathTypes' with matrices
     [5, 6, 7, 8]
   ]
 
+  const efg: MathArray = [1, 2, 3, 4, 5]
+  const fgh: MathArray = [2, 3, 4, 5, 6]
+
   const Mbcd = math.matrix(bcd)
   const Mabc = math.matrix(abc)
 
@@ -1462,8 +1514,11 @@ Math types examples: Type results after multiplying  'MathTypes' with matrices
   const _r2 = math.multiply(a, b)
 
   // 1D JS Array
-  const r3 = math.multiply(abc, bcd)
-  const _r31 = r3[1] // By default least promised valid syntax
+  const r3 = math.multiply(abc, bcd) // 1D * 2D => Array
+  const r3a = math.multiply(efg, fgh) // 1D * 1D => Scalar
+
+  const _r31 = r3[1]
+  assert.strictEqual(typeof r3a, 'number')
 
   // 2D JS Array
   const r12 = math.multiply(bcd, bcd)
@@ -1473,7 +1528,8 @@ Math types examples: Type results after multiplying  'MathTypes' with matrices
   if (multiDimensional(r12)) {
     const _r1211 = r12[1][1]
   }
-  const _r121 = r12[1] // Valid syntax
+
+  const _r121 = r12[1]
 
   // Matrix: matrix * vector
   const r7 = math.multiply(Mabc, bcd)
@@ -1777,6 +1833,54 @@ Function ceil examples
     math.complex(3.3, -2.7)
   )
 
+  // unit input
+  const u1 = math.unit(3.2, 'cm')
+  const u2 = math.unit('cm')
+  const u3 = math.unit(5.51, 'cm')
+
+  // unit array input
+  const unitArray: MathArray<Unit> = [u1, u3]
+  const array = [u1, u3, 1]
+  array.pop()
+  const array2 = [
+    [u1, u3],
+    [1, 5]
+  ]
+  array2.pop()
+
+  assert.deepStrictEqual(math.ceil(u1, u2), math.unit(4, 'cm'))
+  assert.deepStrictEqual(math.ceil(u1, 1, u2), math.unit(3.2, 'cm'))
+  assert.deepStrictEqual(math.ceil(unitArray, 1, math.unit('cm')), [
+    math.unit(3.2, 'cm'),
+    math.unit(5.6, 'cm')
+  ])
+
+  // Can assert that the array is a Unit[]
+  assert.deepStrictEqual(math.ceil(array as Unit[], 1, math.unit('cm')), [
+    math.unit(3.2, 'cm'),
+    math.unit(5.6, 'cm')
+  ])
+
+  // Can assert that the array is a Unit[][]
+  assert.deepStrictEqual(math.ceil(array2 as Unit[][], 1, math.unit('cm')), [
+    [math.unit(3.2, 'cm'), math.unit(5.6, 'cm')]
+  ])
+
+  // unit matrix input
+  const unitMatrix = math.matrix<Unit>(unitArray)
+  const matrix = math.matrix([u1, u3])
+
+  assert.deepStrictEqual(
+    math.ceil(unitMatrix, 1, math.unit('cm')),
+    math.matrix([math.unit(3.2, 'cm'), math.unit(5.6, 'cm')])
+  )
+
+  // Can assert that the matrix is a Matrix<Unit>
+  assert.deepStrictEqual(
+    math.ceil(matrix as Matrix<Unit>, 1, math.unit('cm')),
+    math.matrix([math.unit(3.2, 'cm'), math.unit(5.6, 'cm')])
+  )
+
   // array input
   assert.deepStrictEqual(math.ceil([3.2, 3.8, -4.7]), [4, 4, -4])
   assert.deepStrictEqual(math.ceil([3.21, 3.82, -4.71], 1), [3.3, 3.9, -4.7])
@@ -1853,6 +1957,23 @@ Function fix examples
     math.complex(3.2, -2.7)
   )
 
+  // unit input
+  const u1 = math.unit(3.2, 'cm')
+  const u2 = math.unit('cm')
+  const u3 = math.unit(5.51, 'cm')
+  const unitArray = [u1, u3]
+  const unitMatrix = math.matrix<Unit>(unitArray)
+  assert.deepStrictEqual(math.fix(u1, u2), math.unit(3, 'cm'))
+  assert.deepStrictEqual(math.fix(u1, 1, u2), math.unit(3.2, 'cm'))
+  assert.deepStrictEqual(math.fix(unitArray, 1, math.unit('cm')), [
+    math.unit(3.2, 'cm'),
+    math.unit(5.5, 'cm')
+  ])
+  assert.deepStrictEqual(
+    math.fix(unitMatrix, 1, math.unit('cm')),
+    math.matrix([math.unit(3.2, 'cm'), math.unit(5.5, 'cm')])
+  )
+
   // array input
   assert.deepStrictEqual(math.fix([3.2, 3.8, -4.7]), [3, 3, -4])
   assert.deepStrictEqual(math.fix([3.21, 3.82, -4.71], 1), [3.2, 3.8, -4.7])
@@ -1927,6 +2048,23 @@ Function floor examples
   assert.deepStrictEqual(
     math.floor(c, math.bignumber(1)),
     math.complex(3.2, -2.8)
+  )
+
+  // unit input
+  const u1 = math.unit(3.2, 'cm')
+  const u2 = math.unit('cm')
+  const u3 = math.unit(5.51, 'cm')
+  const unitArray = [u1, u3]
+  const unitMatrix = math.matrix<Unit>(unitArray)
+  assert.deepStrictEqual(math.floor(u1, u2), math.unit(3, 'cm'))
+  assert.deepStrictEqual(math.floor(u1, 1, u2), math.unit(3.2, 'cm'))
+  assert.deepStrictEqual(math.floor(unitArray, 1, math.unit('cm')), [
+    math.unit(3.2, 'cm'),
+    math.unit(5.5, 'cm')
+  ])
+  assert.deepStrictEqual(
+    math.floor(unitMatrix, 1, math.unit('cm')),
+    math.matrix([math.unit(3.2, 'cm'), math.unit(5.5, 'cm')])
   )
 
   // array input
@@ -2006,13 +2144,28 @@ Function round examples
   )
 
   // unit input
+  const u1 = math.unit(3.2, 'cm')
+  const u2 = math.unit('cm')
+  const u3 = math.unit(5.51, 'cm')
+  const unitArray = [u1, u3]
+  const unitMatrix = math.matrix<Unit>(unitArray)
+  assert.deepStrictEqual(math.round(u1, u2), math.unit(3, 'cm'))
+  assert.deepStrictEqual(math.round(u1, 1, u2), math.unit(3.2, 'cm'))
   assert.deepStrictEqual(
-    math.round(math.unit('5.21 cm'), math.unit('cm')),
-    math.unit('5 cm')
+    math.round(u1, math.bignumber(1), u2),
+    math.unit(3.2, 'cm')
+  )
+  assert.deepStrictEqual(math.round(unitArray, 1, math.unit('cm')), [
+    math.unit(3.2, 'cm'),
+    math.unit(5.5, 'cm')
+  ])
+  assert.deepStrictEqual(
+    math.round(unitArray, math.bignumber(1), math.unit('cm')),
+    [math.unit(3.2, 'cm'), math.unit(5.5, 'cm')]
   )
   assert.deepStrictEqual(
-    math.round(math.unit('5.21 cm'), 1, math.unit('cm')),
-    math.unit('5.2 cm')
+    math.round(unitMatrix, 1, math.unit('cm')),
+    math.matrix([math.unit(3.2, 'cm'), math.unit(5.5, 'cm')])
   )
 
   // array input
@@ -2694,4 +2847,56 @@ Match types of exact positional arguments.
   const e = math.mode(mathCollection)
   expectTypeOf(e).toMatchTypeOf<MathScalarType[]>()
   assert.deepStrictEqual(e, [1])
+}
+
+/**
+ * N-dimensional array examples
+ */
+{
+  const math = create(all, {})
+
+  const array1 = [1, 2, 3]
+  const array2 = [
+    [1, 2],
+    [3, 4]
+  ]
+  const array3 = [
+    [
+      [1, 2],
+      [3, 4]
+    ],
+    [
+      [5, 6],
+      [7, 8]
+    ]
+  ]
+  const array4 = [
+    [[[1, 2]], [[3, 4]]],
+    [[[5, 6]], [[7, 8]]],
+    [[[9, 10]], [[11, 12]]]
+  ]
+
+  const mixArray3 = [
+    [
+      [1, math.unit(2, 'cm'), math.bignumber(1), math.complex(1, 2)],
+      [3, math.unit(4, 'cm'), math.bignumber(2), math.complex(3, 4)]
+    ],
+    [
+      [5, math.unit(6, 'cm'), math.bignumber(3), math.complex(5, 6)],
+      [7, math.unit(8, 'cm'), math.bignumber(4), math.complex(7, 8)]
+    ]
+  ]
+
+  const unitArray3 = [
+    [[math.unit(1, 'cm'), math.unit(2, 'cm')]],
+    [[math.unit(3, 'cm'), math.unit(4, 'cm')]]
+  ]
+
+  expectTypeOf(array1).toMatchTypeOf<MathArray>()
+  expectTypeOf(array2).toMatchTypeOf<MathArray>()
+  expectTypeOf(array3).toMatchTypeOf<MathArray>()
+  expectTypeOf(array4).toMatchTypeOf<MathArray>()
+
+  expectTypeOf(mixArray3).toMatchTypeOf<MathArray<MathScalarType>>()
+  expectTypeOf(unitArray3).toMatchTypeOf<MathArray<Unit>>()
 }